--- resources/js/pages/Prestamos/Devoluciones.tsx
@@ -1,362 +1,0 @@
-import { useState, useEffect } from 'react';
-import { Head, router, usePage } from '@inertiajs/react';
-import { Search, CheckCircle, AlertCircle, X } from 'lucide-react';
-import AppLayout from '@/layouts/app-layout';
-import { type BreadcrumbItem } from '@/types';
-
-// Tipos
-type Prestamo = {
-  id: number;
-  ejemplar: {
-    id: number;
-    codigo: string;
-    numEjemplar: number;
-    libro: {
-      id: number;
-      titulo: string;
-      isbn: string;
-    };
-  };
-  fecha_prestamo: string;
-  fecha_devolucion: string;
-  estado: string;
-};
-
-type DevolucionesPageProps = {
-  auth: any;
-  prestamos?: Prestamo[];
-  flash?: {
-    success?: string;
-    error?: string;
-  };
-};
-
-// Constantes
-const breadcrumbs: BreadcrumbItem[] = [
-  {
-    title: 'Préstamos',
-    href: '/prestamos',
-  },
-  {
-    title: 'Devoluciones',
-    href: '/prestamos/devoluciones',
-  },
-];
-
-function AlertNotification({
-  type,
-  message,
-  className = '',
-  autoClose = true,
-  duration = 4000,
-}: {
-  type: 'success' | 'error';
-  message: string;
-  className?: string;
-  autoClose?: boolean;
-  duration?: number;
-}) {
-  const [isVisible, setIsVisible] = useState(true);
-  const [animateOut, setAnimateOut] = useState(false);
-
-  useEffect(() => {
-    if (autoClose && message) {
-      const timer = setTimeout(() => {
-        setAnimateOut(true);
-        const hideTimer = setTimeout(() => {
-          setIsVisible(false);
-        }, 500);
-        return () => clearTimeout(hideTimer);
-      }, duration);
-      return () => clearTimeout(timer);
-    }
-  }, [autoClose, duration, message]);
-
-  if (!isVisible || !message) return null;
-
-  const colors = {
-    success: {
-      light: { bg: 'bg-green-100', border: 'border-green-500', text: 'text-green-800', icon: 'text-green-500' },
-      dark: { bg: 'dark:bg-green-800/40', border: 'dark:border-green-500', text: 'dark:text-green-100', icon: 'dark:text-green-400' }
-    },
-    error: {
-      light: { bg: 'bg-red-100', border: 'border-red-500', text: 'text-red-800', icon: 'text-red-500' },
-      dark: { bg: 'dark:bg-red-800/40', border: 'dark:border-red-500', text: 'dark:text-red-100', icon: 'dark:text-red-400' }
-    }
-  };
-
-  const Icon = type === 'success' ? CheckCircle : AlertCircle;
-
-  return (
-    <div className={`fixed top-6 right-6 z-50 ${animateOut ? 'opacity-0 translate-x-20' : 'opacity-100 translate-x-0'} transition-all duration-500 ease-in-out transform ${className}`}>
-      <div
-        className={`max-w-md rounded-lg shadow-xl border-l-4 
-                    ${colors[type].light.border} ${colors[type].dark.border}
-                    ${colors[type].light.bg} ${colors[type].dark.bg} 
-                    flex items-start p-5 transition-all duration-300 animate-slide-in-right`}
-      >
-        <Icon className={`h-6 w-6 mt-0.5 mr-4 flex-shrink-0 ${colors[type].light.icon} ${colors[type].dark.icon}`} />
-        <div className="flex-grow">
-          <p className={`text-base font-semibold ${colors[type].light.text} ${colors[type].dark.text}`}>
-            {message}
-          </p>
-        </div>
-        <button
-          onClick={() => setAnimateOut(true)}
-          className="ml-4 flex-shrink-0 text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200 focus:outline-none"
-        >
-          <X className="h-5 w-5" />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-export default function Devoluciones({
-  auth,
-  prestamos = [],
-  flash = {},
-}: DevolucionesPageProps) {
-  const { errors = {} } = usePage().props as any;
-  const [searchTerm, setSearchTerm] = useState('');
-  const [prestamosFiltrados, setPrestamosFiltrados] = useState<Prestamo[]>([]);
-  const [alerts, setAlerts] = useState<{
-    success: string | null;
-    error: string | null;
-    timestamp: number;
-  }>({
-    success: null,
-    error: null,
-    timestamp: 0
-  });
-
-  // Mostrar notificaciones de flash
-  useEffect(() => {
-    if (flash?.success) {
-      setAlerts(prev => ({
-        ...prev,
-        success: flash.success || null,
-        timestamp: Date.now()
-      }));
-    } else if (flash?.error) {
-      setAlerts(prev => ({
-        ...prev,
-        error: flash.error || null,
-        timestamp: Date.now()
-      }));
-    }
-  }, [flash]);
-
-  // Buscar préstamos por código de estudiante
-  const handleSearch = () => {
-    if (!searchTerm.trim()) return;
-
-    router.get(
-      route('prestamos.buscar'),
-      { codigo_estudiante: searchTerm },
-      {
-        preserveState: true,
-        onSuccess: (page) => {
-          const prestamosActivos = page.props.prestamos || [];
-          setPrestamosFiltrados(prestamosActivos as Prestamo[]);
-
-          if (!Array.isArray(prestamosActivos) || prestamosActivos.length === 0) {
-            setAlerts({
-              success: null,
-              error: 'No se encontraron préstamos activos para este estudiante',
-              timestamp: Date.now()
-            });
-          }
-        },
-        onError: () => {
-          setAlerts({
-            success: null,
-            error: 'Error al buscar los préstamos',
-            timestamp: Date.now()
-          });
-        }
-      }
-    );
-  };
-
-  // Manejar la tecla Enter en el campo de búsqueda
-  const handleKeyPress = (e: React.KeyboardEvent) => {
-    if (e.key === 'Enter') {
-      handleSearch();
-    }
-  };
-
-  const renderAlerts = () => {
-    return (
-      <>
-        {alerts.success && (
-          <AlertNotification
-            key={`success-${alerts.timestamp}`}
-            type="success"
-            message={alerts.success}
-          />
-        )}
-        {alerts.error && (
-          <AlertNotification
-            key={`error-${alerts.timestamp}`}
-            type="error"
-            message={alerts.error}
-          />
-        )}
-      </>
-    );
-  };
-
-  const content = (
-    <div className="py-8 px-6 bg-slate-50 dark:bg-black min-h-screen">
-      {renderAlerts()}
-      <div className="absolute inset-0 overflow-hidden pointer-events-none fixed">
-        <div className="absolute top-1/4 left-1/4 w-96 h-96 rounded-full bg-blue-500/5 blur-3xl dark:bg-blue-600/10"></div>
-        <div className="absolute bottom-1/3 right-1/4 w-80 h-80 rounded-full bg-indigo-500/5 blur-3xl dark:bg-indigo-600/10"></div>
-      </div>
-      <div className="max-w-7xl mx-auto relative z-10">
-        <div className="flex justify-between items-center mb-8">
-          <h1 className="text-3xl font-bold text-gray-900 dark:text-white bg-gradient-to-r from-blue-600 to-indigo-600 bg-clip-text text-transparent">
-            Devoluciones de Libros
-          </h1>
-          <div className="flex gap-4">
-            <div className="relative">
-              <input
-                type="text"
-                value={searchTerm}
-                onChange={(e) => setSearchTerm(e.target.value)}
-                onKeyDown={handleKeyPress}
-                placeholder="Ingrese el código del estudiante"
-                className="w-80 pl-10 py-2.5 pr-4 rounded-lg bg-white dark:bg-gray-800 border border-gray-300 dark:border-gray-700 
-                          text-gray-900 dark:text-white focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500
-                          shadow-sm transition-all duration-200"
-              />
-              <Search className="w-5 h-5 text-gray-400 absolute left-3 top-1/2 transform -translate-y-1/2" />
-            </div>
-            <button
-              onClick={handleSearch}
-              disabled={!searchTerm.trim()}
-              className="px-6 py-2.5 bg-blue-600 hover:bg-blue-700 dark:bg-blue-500 dark:hover:bg-blue-600 
-                       disabled:bg-gray-400 dark:disabled:bg-gray-600 disabled:cursor-not-allowed
-                       text-white font-medium rounded-lg transition-all duration-200 
-                       flex items-center gap-2 shadow-sm hover:shadow-md"
-            >
-              <Search className="w-5 h-5" />
-              Buscar
-            </button>
-          </div>
-        </div>
-
-        {/* Lista de préstamos */}
-        {prestamosFiltrados.length > 0 && (
-          <div className="overflow-hidden rounded-xl border border-gray-200 dark:border-gray-700 bg-white dark:bg-gray-800 shadow-xl">
-            <div className="px-6 py-4 bg-gray-50 dark:bg-gray-700 border-b border-gray-200 dark:border-gray-600">
-              <h3 className="text-lg font-semibold text-gray-900 dark:text-white">
-                Préstamos Activos
-              </h3>
-              <p className="text-sm text-gray-600 dark:text-gray-400 mt-1">
-                {prestamosFiltrados.length} préstamo{prestamosFiltrados.length !== 1 ? 's' : ''} encontrado{prestamosFiltrados.length !== 1 ? 's' : ''}
-              </p>
-            </div>
-            <div className="overflow-x-auto">
-              <table className="w-full">
-                <thead>
-                  <tr className="bg-gray-50 dark:bg-gray-700">
-                    <th className="px-6 py-4 text-left text-xs font-semibold text-gray-500 dark:text-gray-300 uppercase tracking-wider border-b border-gray-200 dark:border-gray-600">
-                      Libro
-                    </th>
-                    <th className="px-6 py-4 text-left text-xs font-semibold text-gray-500 dark:text-gray-300 uppercase tracking-wider border-b border-gray-200 dark:border-gray-600">
-                      Ejemplar
-                    </th>
-                    <th className="px-6 py-4 text-left text-xs font-semibold text-gray-500 dark:text-gray-300 uppercase tracking-wider border-b border-gray-200 dark:border-gray-600">
-                      Fecha Préstamo
-                    </th>
-                    <th className="px-6 py-4 text-left text-xs font-semibold text-gray-500 dark:text-gray-300 uppercase tracking-wider border-b border-gray-200 dark:border-gray-600">
-                      Fecha Devolución
-                    </th>
-                    <th className="px-6 py-4 text-left text-xs font-semibold text-gray-500 dark:text-gray-300 uppercase tracking-wider border-b border-gray-200 dark:border-gray-600">
-                      Estado
-                    </th>
-                  </tr>
-                </thead>
-                <tbody className="divide-y divide-gray-200 dark:divide-gray-700">
-                  {prestamosFiltrados.map((prestamo) => (
-                    <tr key={prestamo.id} className="hover:bg-gray-50 dark:hover:bg-gray-700/50 transition-colors">
-                      <td className="px-6 py-4">
-                        <div>
-                          <div className="font-medium text-gray-900 dark:text-white mb-1">
-                            {prestamo.ejemplar.libro.titulo}
-                          </div>
-                          <div className="text-sm text-gray-600 dark:text-gray-400">
-                            ISBN: {prestamo.ejemplar.libro.isbn}
-                          </div>
-                        </div>
-                      </td>
-                      <td className="px-6 py-4">
-                        <div className="text-sm font-medium text-gray-900 dark:text-white">
-                          #{prestamo.ejemplar.numEjemplar}
-                        </div>
-                        <div className="text-sm text-gray-600 dark:text-gray-400">
-                          {prestamo.ejemplar.codigo}
-                        </div>
-                      </td>
-                      <td className="px-6 py-4 text-sm text-gray-700 dark:text-gray-300">
-                        {new Date(prestamo.fecha_prestamo).toLocaleDateString('es-ES', {
-                          year: 'numeric',
-                          month: 'short',
-                          day: 'numeric'
-                        })}
-                      </td>
-                      <td className="px-6 py-4 text-sm text-gray-700 dark:text-gray-300">
-                        {new Date(prestamo.fecha_devolucion).toLocaleDateString('es-ES', {
-                          year: 'numeric',
-                          month: 'short',
-                          day: 'numeric'
-                        })}
-                      </td>
-                      <td className="px-6 py-4">
-                        <span className={`inline-flex items-center px-3 py-1 text-xs font-medium rounded-full ${
-                          prestamo.estado.toLowerCase() === 'activo'
-                            ? 'bg-amber-100 dark:bg-amber-900/50 text-amber-800 dark:text-amber-200'
-                            : prestamo.estado.toLowerCase() === 'vencido'
-                            ? 'bg-red-100 dark:bg-red-900/50 text-red-800 dark:text-red-200'
-                            : 'bg-gray-100 dark:bg-gray-800 text-gray-800 dark:text-gray-200'
-                        }`}>
-                          {prestamo.estado}
-                        </span>
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          </div>
-        )}
-
-        {/* Estado vacío cuando no hay resultados */}
-        {searchTerm && prestamosFiltrados.length === 0 && !alerts.error && (
-          <div className="overflow-hidden rounded-xl border border-gray-200 dark:border-gray-700 bg-white dark:bg-gray-800 shadow-xl">
-            <div className="p-12 text-center">
-              <div className="w-16 h-16 mx-auto mb-4 bg-gray-100 dark:bg-gray-700 rounded-full flex items-center justify-center">
-                <Search className="w-8 h-8 text-gray-400 dark:text-gray-500" />
-              </div>
-              <h3 className="text-lg font-medium text-gray-900 dark:text-white mb-2">
-                No se encontraron préstamos
-              </h3>
-              <p className="text-gray-600 dark:text-gray-400">
-                No hay préstamos activos para el código de estudiante ingresado.
-              </p>
-            </div>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-
-  return (
-    <AppLayout breadcrumbs={breadcrumbs}>
-      <Head title="Devoluciones de Libros" />
-      {content}
-    </AppLayout>
-  );
-}
